+++ modelcraft/jobs/refmac.py
@@ -16,6 +16,8 @@ class RefmacResult:
     rwork: float
     rfree: float
     fsc: float
+    data_completeness: float
+    resolution_high: float
     seconds: float
 
 
@@ -60,6 +62,8 @@ class _Refmac(Job):
             rwork=float(rworks[-1].text),
             rfree=float(rfrees[-1].text),
             fsc=float(fscs[-1].text),
+            data_completeness=float(xml.find("Overall_stats/data_completeness").text),
+            resolution_high=float(xml.find("Overall_stats/resolution_high").text),
             seconds=self._seconds,
         )