+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.  
 This project adheres to [Semantic Versioning](https://semver.org).
 
+## [2.1.0] - 2022-01-03
+
+### Added
+
+- Phase match job and tests
+- Refmac resolution and completeness
+
+### Fixed
+
+- Termination after Buccaneer builds no residues
+
 ## [2.0.5] - 2021-12-03
 
 ### Added